+++ sdk/monitor/opentelemetry-exporter-azure-monitor/review/opentelemetry-exporter-azure-monitor.api.md
@@ -5,79 +5,41 @@
 ```ts
 
 import { ExportResult } from '@opentelemetry/core';
-import { Logger } from '@opentelemetry/api';
 import { ReadableSpan } from '@opentelemetry/tracing';
 import { SpanExporter } from '@opentelemetry/tracing';
 
 // @public (undocumented)
 export interface AzureExporterConfig {
     // (undocumented)
-    batchSendRetryIntervalMs: number;
-    // (undocumented)
     connectionString?: string;
-    // (undocumented)
-    endpointUrl: string;
-    // (undocumented)
-    instrumentationKey: string;
-    // (undocumented)
-    logger?: Logger;
-    // (undocumented)
-    maxConsecutiveFailuresBeforeWarning: number;
 }
 
 // @public (undocumented)
-export abstract class AzureMonitorBaseExporter implements BaseExporter {
-    constructor(_options?: Partial<AzureExporterConfig>);
+export class AzureMonitorTraceExporter implements SpanExporter {
+    constructor(options?: Partial<AzureExporterConfig>);
     // (undocumented)
     addTelemetryProcessor(processor: TelemetryProcessor): void;
     // (undocumented)
-    protected _applyTelemetryProcessors(envelopes: TelemetryItem[]): TelemetryItem[];
-    // (undocumented)
     clearTelemetryProcessors(): void;
-    // (undocumented)
-    exportEnvelopes(payload: TelemetryItem[]): Promise<ExportResult>;
-    // (undocumented)
-    protected readonly _logger: Logger;
-    // (undocumented)
-    protected readonly _options: AzureExporterConfig;
-    // (undocumented)
-    protected readonly _persister: PersistentStorage;
-    // (undocumented)
-    protected _retryTimer: NodeJS.Timer | null;
-    // (undocumented)
-    protected readonly _sender: Sender;
-    // (undocumented)
-    shutdown(): void;
-    // (undocumented)
-    protected _telemetryProcessors: TelemetryProcessor[];
-}
-
-// @public (undocumented)
-export class AzureMonitorTraceExporter extends AzureMonitorBaseExporter implements SpanExporter {
-    constructor(options?: Partial<AzureExporterConfig>);
     // (undocumented)
     export(spans: ReadableSpan[], resultCallback: (result: ExportResult) => void): Promise<void>;
     // (undocumented)
     shutdown(): void;
-}
-
-// @public (undocumented)
-export interface BaseExporter {
-    // (undocumented)
-    addTelemetryProcessor(processor: TelemetryProcessor): void;
-    // (undocumented)
-    clearTelemetryProcessors(): void;
-    // (undocumented)
-    exportEnvelopes(envelopes: TelemetryItem[]): Promise<ExportResult>;
-}
+    }
 
 // @public
-export type ContextTagKeys = "ai.application.ver" | "ai.device.id" | "ai.device.locale" | "ai.device.model" | "ai.device.oemName" | "ai.device.osVersion" | "ai.device.type" | "ai.location.ip" | "ai.location.country" | "ai.location.province" | "ai.location.city" | "ai.operation.id" | "ai.operation.name" | "ai.operation.parentId" | "ai.operation.syntheticSource" | "ai.operation.correlationVector" | "ai.session.id" | "ai.session.isFirst" | "ai.user.accountId" | "ai.user.id" | "ai.user.authUserId" | "ai.cloud.role" | "ai.cloud.roleVer" | "ai.cloud.roleInstance" | "ai.cloud.location" | "ai.internal.sdkVersion" | "ai.internal.agentVersion" | "ai.internal.nodeName" | string;
-
-// @public (undocumented)
-export type Measurements = {
-    [key: string]: number;
-};
+export interface Envelope {
+    data?: MonitorBase;
+    instrumentationKey?: string;
+    name: string;
+    sampleRate?: number;
+    sequence?: string;
+    tags?: {
+        [propertyName: string]: string;
+    };
+    time: Date;
+    version?: number;
+}
 
 // @public
 export interface MonitorBase {
@@ -92,64 +54,7 @@ export interface MonitorDomain {
 }
 
 // @public (undocumented)
-export interface MSLink {
-    // (undocumented)
-    id: string;
-    // (undocumented)
-    operation_Id: string;
-}
-
-// @public (undocumented)
-export interface PersistentStorage {
-    // (undocumented)
-    push(value: unknown[]): Promise<boolean>;
-    // (undocumented)
-    shift(): Promise<unknown>;
-}
-
-// @public (undocumented)
-export type Properties = {
-    [key: string]: Properties | PropertyType;
-};
-
-// @public (undocumented)
-export type PropertyType = string | number | boolean | object | Array<PropertyType>;
-
-// @public (undocumented)
-export interface Sender {
-    // (undocumented)
-    send(payload: unknown[]): Promise<SenderResult>;
-    // (undocumented)
-    shutdown(): void;
-}
-
-// @public (undocumented)
-export type SenderResult = {
-    statusCode: number;
-    result: string;
-};
-
-// @public (undocumented)
-export type Tags = {
-    [key in ContextTagKeys]: string;
-};
-
-// @public
-export interface TelemetryItem {
-    data?: MonitorBase;
-    instrumentationKey?: string;
-    name: string;
-    sampleRate?: number;
-    sequence?: string;
-    tags?: {
-        [propertyName: string]: string;
-    };
-    time: Date;
-    version?: number;
-}
-
-// @public (undocumented)
-export type TelemetryProcessor = (envelope: TelemetryItem) => boolean | void;
+export type TelemetryProcessor = (envelope: Envelope) => boolean | void;
 
 
 // (No @packageDocumentation comment for this package)
