+++ sdk/monitor/opentelemetry-exporter-azure-monitor/src/index.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from "./export/trace";
-export * from "./export/exporter";
-export * from "./types";
-export { TelemetryItem, ContextTagKeys, MonitorBase, MonitorDomain } from "./generated";
+export { AzureMonitorTraceExporter } from "./export/trace";
+export { TelemetryProcessor } from "./types";
+export { TelemetryItem as Envelope, MonitorBase, MonitorDomain } from "./generated";
 export { AzureExporterConfig } from "./config";
