+++ sdk/monitor/opentelemetry-exporter-azure-monitor/src/types.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ExportResult } from "@opentelemetry/core";
 import { ContextTagKeys, TelemetryItem as Envelope } from "./generated";
 
 export type Tags = { [key in ContextTagKeys]: string };
@@ -14,12 +13,6 @@ export interface MSLink {
 export type Measurements = { [key: string]: number };
 export type TelemetryProcessor = (envelope: Envelope) => boolean | void;
 export type SenderResult = { statusCode: number; result: string };
-
-export interface BaseExporter {
-  addTelemetryProcessor(processor: TelemetryProcessor): void;
-  clearTelemetryProcessors(): void;
-  exportEnvelopes(envelopes: Envelope[]): Promise<ExportResult>;
-}
 
 export interface Sender {
   send(payload: unknown[]): Promise<SenderResult>;
